--- a/plumbing/pool.go
+++ b/plumbing/pool.go
@@ -4,11 +4,10 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"math/rand"
 	"sync"
-	"sync/atomic"
 	"time"
-	"unsafe"
+
+	throughputlb "code.cloudfoundry.org/grpc-throughputlb"
 
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
@@ -18,8 +17,9 @@ type Pool struct {
 	size int
 
 	mu       sync.RWMutex
-	dopplers map[string][]unsafe.Pointer
+	dopplers map[string]clientInfo
 	dialOpts []grpc.DialOption
+	numConns int
 }
 
 type clientInfo struct {
@@ -27,110 +27,80 @@ type clientInfo struct {
 	closer io.Closer
 }
 
-func NewPool(size int, opts ...grpc.DialOption) *Pool {
+func NewPool(numConns int, opts ...grpc.DialOption) *Pool {
 	return &Pool{
-		size:     size,
-		dopplers: make(map[string][]unsafe.Pointer),
+		dopplers: make(map[string]clientInfo),
 		dialOpts: opts,
+		numConns: numConns,
 	}
 }
 
 func (p *Pool) RegisterDoppler(addr string) {
-	clients := make([]unsafe.Pointer, p.size)
-
-	p.mu.Lock()
-	p.dopplers[addr] = clients
-	p.mu.Unlock()
-
-	for i := 0; i < p.size; i++ {
-		go p.connectToDoppler(addr, clients, i)
-	}
+	go p.connectToDoppler(addr)
 }
 
 func (p *Pool) Subscribe(dopplerAddr string, ctx context.Context, req *SubscriptionRequest) (Doppler_BatchSubscribeClient, error) {
 	p.mu.RLock()
-	clients := p.dopplers[dopplerAddr]
+	ci, ok := p.dopplers[dopplerAddr]
 	p.mu.RUnlock()
 
-	client := p.fetchClient(clients)
-
-	if client == nil {
+	if !ok {
 		return nil, fmt.Errorf("no connections available for subscription")
 	}
 
-	return client.BatchSubscribe(ctx, req)
+	return ci.client.BatchSubscribe(ctx, req)
 }
 
 func (p *Pool) ContainerMetrics(dopplerAddr string, ctx context.Context, req *ContainerMetricsRequest) (*ContainerMetricsResponse, error) {
 	p.mu.RLock()
-	clients := p.dopplers[dopplerAddr]
+	ci, ok := p.dopplers[dopplerAddr]
 	p.mu.RUnlock()
 
-	client := p.fetchClient(clients)
-
-	if client == nil {
+	if !ok {
 		return nil, fmt.Errorf("no connections available for container metrics")
 	}
 
-	return client.ContainerMetrics(ctx, req)
+	return ci.client.ContainerMetrics(ctx, req)
 }
 
 func (p *Pool) RecentLogs(dopplerAddr string, ctx context.Context, req *RecentLogsRequest) (*RecentLogsResponse, error) {
 	p.mu.RLock()
-	clients := p.dopplers[dopplerAddr]
+	ci, ok := p.dopplers[dopplerAddr]
 	p.mu.RUnlock()
 
-	client := p.fetchClient(clients)
-
-	if client == nil {
+	if !ok {
 		return nil, fmt.Errorf("no connections available for recent logs")
 	}
 
-	return client.RecentLogs(ctx, req)
+	return ci.client.RecentLogs(ctx, req)
 }
 
 func (p *Pool) Close(dopplerAddr string) {
 	p.mu.Lock()
-	clients := p.dopplers[dopplerAddr]
+	ci, ok := p.dopplers[dopplerAddr]
 	delete(p.dopplers, dopplerAddr)
 	p.mu.Unlock()
 
-	for i := range clients {
-		clt := atomic.LoadPointer(&clients[i])
-		if clt == nil ||
-			(*clientInfo)(clt) == nil {
-			continue
-		}
-
-		client := *(*clientInfo)(clt)
-		client.closer.Close()
+	if ok {
+		ci.closer.Close()
 	}
 }
 
-func (p *Pool) fetchClient(clients []unsafe.Pointer) DopplerClient {
-	seed := rand.Int()
-	for i := range clients {
-		idx := (i + seed) % p.size
-		clt := atomic.LoadPointer(&clients[idx])
-		if clt == nil ||
-			(*clientInfo)(clt) == nil {
-			continue
-		}
-
-		client := *(*clientInfo)(clt)
-		return client.client
+func (p *Pool) connectToDoppler(addr string) {
+	opts := []grpc.DialOption{
+		grpc.WithBalancer(
+			throughputlb.NewThroughputLoadBalancer(100, p.numConns),
+		),
 	}
 
-	return nil
-}
+	for _, o := range p.dialOpts {
+		opts = append(opts, o)
+	}
 
-func (p *Pool) connectToDoppler(addr string, clients []unsafe.Pointer, idx int) {
 	for {
-		// TODO: Do we need this log line? It prints the same log line based
-		// on pool size
 		log.Printf("adding doppler %s", addr)
 
-		conn, err := grpc.Dial(addr, p.dialOpts...)
+		conn, err := grpc.Dial(addr, opts...)
 		if err != nil {
 			// TODO: We don't yet understand how this could happen, we should.
 			// TODO: Replace with exponential backoff.
@@ -139,13 +109,13 @@ func (p *Pool) connectToDoppler(addr string, clients []unsafe.Pointer, idx int) 
 			continue
 		}
 
-		client := NewDopplerClient(conn)
-		info := clientInfo{
-			client: client,
+		p.mu.Lock()
+		p.dopplers[addr] = clientInfo{
+			client: NewDopplerClient(conn),
 			closer: conn,
 		}
+		p.mu.Unlock()
 
-		atomic.StorePointer(&clients[idx], unsafe.Pointer(&info))
 		return
 	}
 }