@apoydence 2017-11-01
@enocom @bradylove And I are going to deploy a version of Loggregator to an environment where we can load test it. If all goes well, we'll merge it.